# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Blog API - Proyecto Integrador 2
  description: |
    API REST para gestionar autores, posts y comentarios.
    
    ## Cómo usar
    1. Clona el repositorio y copia `.env.example` a `.env`.
    2. Ejecuta `setup.sql` contra tu base de datos PostgreSQL.
    3. Corre `npm install && npm start`.
  version: 1.0.0
  contact:
    name: Proyecto Integrador 2

servers:
  - url: http://localhost:3000
    description: Servidor local
  - url: https://tu-app.up.railway.app
    description: Producción (Railway)

tags:
  - name: Health
  - name: Authors
  - name: Posts
  - name: Comments

paths:
  /health:
    get:
      tags: [Health]
      summary: Estado de la API
      responses:
        '200':
          description: API operativa
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  timestamp: { type: string, format: date-time }

  # ── AUTHORS ──────────────────────────────────────────────────
  /authors:
    get:
      tags: [Authors]
      summary: Listar todos los autores
      responses:
        '200':
          description: Lista de autores
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Author' }
    post:
      tags: [Authors]
      summary: Crear un autor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthorInput' }
      responses:
        '201':
          description: Autor creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Author' }
        '400':
          description: Validación fallida o email duplicado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /authors/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Authors]
      summary: Obtener autor por ID
      responses:
        '200':
          description: Autor encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Author' }
        '404':
          description: No encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [Authors]
      summary: Actualizar autor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthorInput' }
      responses:
        '200':
          description: Autor actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Author' }
        '400':
          description: Validación fallida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: No encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Authors]
      summary: Eliminar autor (también elimina sus posts y comentarios)
      responses:
        '204':
          description: Eliminado correctamente
        '404':
          description: No encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ── POSTS ─────────────────────────────────────────────────────
  /posts:
    get:
      tags: [Posts]
      summary: Listar todos los posts
      responses:
        '200':
          description: Lista de posts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Post' }
    post:
      tags: [Posts]
      summary: Crear un post
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostInput' }
      responses:
        '201':
          description: Post creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '400':
          description: Validación fallida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /posts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Posts]
      summary: Obtener post por ID
      responses:
        '200':
          description: Post encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '404':
          description: No encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [Posts]
      summary: Actualizar post
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostInput' }
      responses:
        '200':
          description: Post actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '400':
          description: Validación fallida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: No encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Posts]
      summary: Eliminar post
      responses:
        '204':
          description: Eliminado correctamente
        '404':
          description: No encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /posts/author/{authorId}:
    parameters:
      - name: authorId
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Posts]
      summary: Listar posts de un autor con detalle del autor
      responses:
        '200':
          description: Posts con información del autor
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PostWithAuthor' }

  # ── COMMENTS ─────────────────────────────────────────────────
  /comments:
    get:
      tags: [Comments]
      summary: Listar comentarios (opcionalmente filtrar por post)
      parameters:
        - name: post_id
          in: query
          schema: { type: integer }
          description: Filtrar comentarios de un post específico
      responses:
        '200':
          description: Lista de comentarios
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Comment' }
    post:
      tags: [Comments]
      summary: Crear un comentario
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommentInput' }
      responses:
        '201':
          description: Comentario creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
        '400':
          description: Validación fallida
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /comments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Comments]
      summary: Obtener comentario por ID
      responses:
        '200':
          description: Comentario encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
        '404':
          description: No encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Comments]
      summary: Eliminar comentario
      responses:
        '204':
          description: Eliminado correctamente
        '404':
          description: No encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

components:
  schemas:
    Author:
      type: object
      properties:
        id:         { type: integer, example: 1 }
        name:       { type: string,  example: "Ana García" }
        email:      { type: string,  format: email, example: "ana@example.com" }
        bio:        { type: string,  nullable: true, example: "Desarrolladora full-stack" }
        created_at: { type: string,  format: date-time }

    AuthorInput:
      type: object
      required: [name, email]
      properties:
        name:  { type: string, example: "Ana García" }
        email: { type: string, format: email, example: "ana@example.com" }
        bio:   { type: string, nullable: true }

    Post:
      type: object
      properties:
        id:         { type: integer, example: 1 }
        title:      { type: string,  example: "Introducción a Node.js" }
        content:    { type: string,  example: "Node.js es un runtime..." }
        author_id:  { type: integer, example: 1 }
        published:  { type: boolean, example: true }
        created_at: { type: string,  format: date-time }

    PostWithAuthor:
      allOf:
        - $ref: '#/components/schemas/Post'
        - type: object
          properties:
            author: { $ref: '#/components/schemas/Author' }

    PostInput:
      type: object
      required: [title, content, author_id]
      properties:
        title:     { type: string,  example: "Mi nuevo post" }
        content:   { type: string,  example: "Contenido del post..." }
        author_id: { type: integer, example: 1 }
        published: { type: boolean, default: false }

    Comment:
      type: object
      properties:
        id:         { type: integer, example: 1 }
        body:       { type: string,  example: "Excelente artículo!" }
        post_id:    { type: integer, example: 1 }
        author_id:  { type: integer, example: 2 }
        created_at: { type: string,  format: date-time }
        author:     { $ref: '#/components/schemas/Author' }

    CommentInput:
      type: object
      required: [body, post_id, author_id]
      properties:
        body:      { type: string,  example: "Excelente artículo!" }
        post_id:   { type: integer, example: 1 }
        author_id: { type: integer, example: 2 }

    Error:
      type: object
      properties:
        error: { type: string, example: "Mensaje de error" }